moved widgets to formality & disabled past dates on datetimepicker.
