+++ public_website/apps/formality/templates/widgets/xdsoft_datetimepicker.html
@@ -4,6 +4,7 @@
 	$(function(){
 		$("input[name='{{ widget.name }}']").datetimepicker({
 			format: 'Y-m-d H:i:00',
+			minDate:0,
 		})
 	});
 </script>
